Remove outdated reference in UFW guide (#3534)
The UFW guide currently says "do not follow the Configure a Firewall section in How to Secure Your Server guide", presumably because the "How to Secure Your Server" guide used to include only instructions for using `iptables`. However, that guide now includes links to guides that provide different options for firewall configuration (including this UFW guide), so this "warning" in the UFW guide no longer makes sense. This commit removes it.